requirements.txt
streamlit==1.29.0
pandas==2.1.1
requests==2.31.0
plotly==5.15.0
python-dotenv==1.1.1
st_autorefresh